fix: move to trusted publishing (#1551)
* fix: move to trusted publishing

Also moves the CI workflow to npm and updates CI in general. Moves the release workflow to a separate one that is called when we want to do a release. This manual action is in line with how we do `@sequelize/*` releases, I have removed the tests from it since tests need to pass before we merge.

* build(deps): update lockfile

* ci: use npm 7 on Node 10